Move MainActivity and TransactionsFragmentClicks into package "ui"
