+++ app/src/main/java/ru/aevd/androidacademymovieapp/ui/MainActivity.kt
@@ -1,7 +1,8 @@
-package ru.aevd.androidacademymovieapp
+package ru.aevd.androidacademymovieapp.ui
 
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
+import ru.aevd.androidacademymovieapp.R
 import ru.aevd.androidacademymovieapp.domain.entities.Movie
 import ru.aevd.androidacademymovieapp.ui.fragments.FragmentMoviesDetails
 import ru.aevd.androidacademymovieapp.ui.fragments.FragmentMoviesList
