+++ app/src/main/res/layout/fragment_movies_details.xml
@@ -11,7 +11,7 @@
         android:id="@+id/layout"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        tools:context=".MainActivity">
+        tools:context=".ui.MainActivity">
 
         <androidx.constraintlayout.widget.Guideline
             android:id="@+id/guideline_top"
